SAK-19035
Created a new value for the "system" attribute of the main document element in the archive export files. The new value is "Sakai 2.8".
Export and merge code now looks for this new value. When it is encountered the site.xml file is processed differently. "roleId" is
now encoded as an attribute on a "role" element instead of using the value of "roleId" as the element name.

Processing of "Sakai 1.0" files remains the same. All processing of "Sakai 2.8" files are protected by conditional statements checking
for that value.


git-svn-id: https://source.sakaiproject.org/svn/common/trunk@82043 66ffb92e-73f9-0310-93c1-f5514f145a0a